Complete telemetry library code
This includes a lot of metrics we probably don't care about and since
this is already a fork of chromite's implementation it would be
preferable to not support metrics we con't actually care about. This is
being uploaded as backup/proof of concept and only the portions we care
about will be added as needed.

Bug: 326277821
Change-Id: I0212b5819e824e8e1287368de47c9ff31bdd30bd