[cpplint] simplify input encoding and fix a bug for stdin
Reading inputs from stdin helps seamless integration of
the lint tools in IDE tools, such as Visual Studio, because
the linter can be performed against the unsaved text lines shown
on the editor.

This CL fixes the following 2 issues.

1) stdin support in git cl lint.
While cpplint supports inputs from stdin, if the filename is -,
git cl lint doesn't support stdin inputs.

2) cpplint raises an exception when reading inputs from stdin
It uses StreamReaderWriter with implicit type conversion between
str and byte, but it raises an exception in Python3.
: TypeError: can't concat str to bytes

The logic also has unnecessary part to strip off '\r' and '\n', but
the stripping is unnecessary with splitlines().

This CL simply replace all the relevant logic with the builtin
open() and splitlines().

Bug: 372288610
Change-Id: Ib2d5a3534ac0c5afcfcac64b708becde41612881