[cpplint] add stdin support in stdin, and fix a missing py2-py3 conv.
Reading inputs from stdin helps seamless integration of
the lint tools in IDE tools, such as Visual Studio, because
the linter can be performed against the unsaved text lines shown
on the editor.

This CL fixes the following 2 issues.
1) stdin support in git cl lint.
While cpplint supports inputs from stdin, if the filename is -,
git cl lint doesn't support stdin inputs.

2) cpplint raises an exception if stdin is given.
- TypeError: can't concat str to bytes

This seems to be misssed in the py2-py3 conversion.


Bug: 372288610
Change-Id: Ib2d5a3534ac0c5afcfcac64b708becde41612881