Add timeout to git call in windows toolchain update script
VSCode will intercept git calls from its embedded terminal under at
least some circumstances. If this happens, it will create a small,
easily-missable dialogue box, causing the script to hang until the
box is closed. As a far as I can tell, there's no way to prevent
VSCode from doing this.

To prevent this from happening during windows toolchain updates,
we add a timeout to the relevant git call, which is checking for
certain permissions. If we don't get a quick response, we'll just
assume the permissions don't exist.

The call in question is one of three redundant checks, and returning
false is the conservative option, so this shouldn't cause any problems.
The only minor annoyance is that this leaves the dialogue box at the
top of the VSCode UI until the user does something about it, but once
the script has continued the box is harmless.

Bug: 376067358
Change-Id: I1a20257c52c3569a03797705d7460feedee9a124