[Depot tools]Add comment and constants
While reading the code, I had a hard time understanding a method
(because I don’t know the gerrit API). So I added documentation with a
link to the relevant fields of the API.

I considered adding a note about the most common fields, but decided
if anyone need to know, they can follow the link.

I also replaced some literals by constant. It’s easier to find where
they are used in the context of the API, and ensure there is no risk
for typo while reusing the same literal.

I considered adding it in gerrit_util, but that would mean it could
not be used in metrics_utils.

Bug: None
Change-Id: Ib536bc95af8e2b366f0c3eeb6f9a5fb69bed662f