[Depot tools]`git cl upload` ignores commit messages changes
Normally, when some revisions are done on gerrit and not known
locally, the user is warned and asked whether they want to pull
revision locally.

However, when the revision consists in a change in the commit message,
this warning is useless. Commit messages are ignored during CL
upload. Pulling the revision is also useless.

So, in the case where all revisions are changes in commit message, the
user is not asked whether they want to apply patchset.

Fixed: 376910768
Change-Id: I8017711a62eea3620ae8022b3b93f94278580870