Do not block CQ on pylint warnings
pylint is too strict for it to block CQ. A lot of times it is better to
leave the code as is than litter warning disables in the code (eg:
adding 30 chars to an 81 char line to disable a check because it was 1
char over the limit). Switch to prompt for upload but allow the dev to
submit the CL anyways.

Change-Id: I8f832bcc2fcfdd1e9854ebd5352e782dcadd450b