[git] Remove tags / branches filter from ls-remote
The flag to filter for branches was renamed in git 2.46.0 [1] from
`-h` to `-b`. This causes previous versions of git to fail when running
ls-remote.

To be compatible across a version upgrade, a second git call would be
required to retrieve the version and determine the correct flag.
However, the filter feature is not used yet and can also be solved by
specifying the absolute ref path within the recipe. Therefore, this
change removes ls-remote's filter flags.

[1] https://git-scm.com/docs/git-ls-remote/2.42.0

R=machenbach@chromium.org, gavinmak@google.com

Bug: none
Change-Id: I840fb3ac56daf47c798d5c32f00de841b8db2771